Remove componentWillMount
`componentWillMount` is deprecated and therefore warnings are shown when using RN 0.60+.

```
Warning: componentWillMount is deprecated and will be removed in the next major version. Use componentDidMount instead. As a temporary workaround, you can rename to UNSAFE_componentWillMount.

Please update the following components: WrappedAnimatedInteractableView
```

The lifecycle isn't used anymore so we can safely remove it.